teeny weeny cleanup (The project is still shit)
